Hooked up checkbox to task DB (#7)
Enabled communication with the DB to dynamically store the completed
state (check box) of a task.

I modified how to handle CORS to allow this communication and allowed
users to create their own credentials.json file in "backend/store/core/"
to load their mysql configurations instead of storing it in the
database.go file.